pyproject.toml
[tool.poetry]
name = "microservices_grpc"
version = "0.1.0"
description = ""
authors = ["vildan-valeev <vildan.maydanovich.valeev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.8.1"
grpcio = "^1.43.0"
grpcio-tools = "^1.43.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
